+++ src/lcm/grids/piecewise.py
@@ -87,9 +87,7 @@ class PiecewiseLinSpacedGrid(ContinuousGrid):
     def get_coordinate(self, value: Array) -> Array: ...
     def get_coordinate(self, value: ScalarFloat | Array) -> ScalarFloat | Array:
         """Return the generalized coordinate of a value in the grid."""
-        piece_idx = jnp.searchsorted(self._breakpoints, value, side="right").astype(
-            jnp.int32
-        )
+        piece_idx = jnp.searchsorted(self._breakpoints, value, side="right")
         local_coord = grid_coordinates.get_linspace_coordinate(
             value=value,
             start=self._piece_starts[piece_idx],
@@ -160,9 +158,7 @@ class PiecewiseLogSpacedGrid(ContinuousGrid):
     def get_coordinate(self, value: Array) -> Array: ...
     def get_coordinate(self, value: ScalarFloat | Array) -> ScalarFloat | Array:
         """Return the generalized coordinate of a value in the grid."""
-        piece_idx = jnp.searchsorted(self._breakpoints, value, side="right").astype(
-            jnp.int32
-        )
+        piece_idx = jnp.searchsorted(self._breakpoints, value, side="right")
         local_coord = grid_coordinates.get_logspace_coordinate(
             value=value,
             start=self._piece_starts[piece_idx],
