+++ src/lcm/simulation/simulate.py
@@ -355,7 +355,7 @@ def _lookup_values_from_indices(
     nd_indices = vmapped_unravel_index(flat_indices, grids_shapes)
     return MappingProxyType(
         {
-            name: grid[index.astype(jnp.int32)]
+            name: grid[index]
             for (name, grid), index in zip(grids.items(), nd_indices, strict=True)
         }
     )
@@ -385,7 +385,7 @@ def _compute_starting_periods(
 
     """
     age_values = jnp.asarray(ages.values)
-    starting_periods = jnp.searchsorted(age_values, initial_ages).astype(jnp.int32)
+    starting_periods = jnp.searchsorted(age_values, initial_ages)
 
     # Clamp indices to valid range before accessing age_values. searchsorted can
     # return len(age_values) for ages beyond the grid maximum.
